requirements.txt
langchain==0.1.13
langchain-google-genai==0.0.12
langchain-core==0.1.30
langgraph==0.0.32
python-dotenv==1.0.0
google-search-results==2.4.2
pydantic==2.5.0
requests==2.31.0
pytest==7.4.3
pytest-asyncio==0.21.1
aiohttp==3.9.1
